# openapi/jewelheart.yaml
openapi: 3.0.3
info:
  title: JewelHeart Admin API
  description: |
    Retreat volunteer scheduling API for JewelHeart mobile apps (iOS/Android).
    Hosted on the same Cloudflare tunnel as KarmaDots: **https://api.karmadots.org**.

    All endpoints except `/jewelheart/health` require a **Firebase ID token** (same as KarmaDots private-server):
    `Authorization: Bearer <firebase_id_token>`.

    **ACL:** The server must verify the token and enforce JewelHeart admin / per-retreat access (implementation detail; not expressed in this contract).

    **Volunteer calendar feeds** (webcal/ICS) and **assignment confirmation** payloads are specified under
    `/jewelheart/calendar-feed/{feedToken}` and `/jewelheart/assignment-confirmations/...` (see tags Calendar, Confirmations).

    **Reminder policy:** native calendar clients apply `VALARM`/alarms from subscribed ICS; do not rely on server push for calendar-specific alerts.
  version: 0.1.0
  license:
    name: Proprietary
servers:
  - url: https://api.karmadots.org
    description: Production (Cloudflare tunnel → private-server)
...
paths:
...
  /jewelheart/assignment-confirmations/{sealedConfirmationToken}:
    parameters:
      - $ref: "#/components/parameters/SealedConfirmationToken"
    get:
      tags: [Confirmations]
      summary: Hosted confirmation splash (tiny HTML fallback)
      description: >
        Serves minimalist HTML referencing two POST targets or inline forms so SMS/email deep-links work without bespoke mobile builds.

        This route is OPTIONAL if deployment instead deep-links straight to an SPA/web property; reserve path for portability.
      operationId: assignmentConfirmationLanding
      security: []
      responses:
        "200":
          description: HTML hint page
          content:
            text/html:
              schema:
                type: string
        "404":
          description: Signature invalid/expired assignment
        "410":
          description: Assignment already finalized or removed
    post:
      tags: [Confirmations]
      summary: Record volunteer intent for an upcoming assignment
      description: >
        Parses `sealedConfirmationToken`, validates MAC/expiry bound to `{assignmentId, volunteerId}`.

        **`committed`** (or omission of response body / separate keep-alive webhook) SHOULD be a durable no-op: assignment remains.

        **`withdrawn`** instructs backend to DELETE the assignment row (preferred) OR enqueue moderator review depending on parish policy —
        MUST be explicit in changelog when choosing non-deleting flows.

        **Push/SMS/email nudges MUST NOT implicitly cancel** when user ignores prompts; unanswered prompts keep DB rows untouched.
      operationId: assignmentConfirmationRespond
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignmentConfirmationRespondRequest"
      responses:
        "200":
          description: Intent persisted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssignmentConfirmationRespondResult"
        "409":
          description: Intent conflicts with staffing policy (capacity / locked retreat)
        "404":
          description: Signature invalid/expired assignment
        "410":
          description: Assignment unavailable
  /jewelheart/retreats/{retreatId}/tasks/{taskId}/assignments:
    parameters:
      - $ref: "#/components/parameters/RetreatId"
      - $ref: "#/components/parameters/TaskId"
    post:
      tags: [Assignments]
      summary: Assign volunteer to task (must be linked to retreat; capacity enforced)
      operationId: createAssignment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignmentCreate"
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assignment"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Over capacity or duplicate assignment
        "401":
          $ref: "#/components/responses/Unauthorized"
  /jewelheart/retreats/{retreatId}/assignments/{assignmentId}:
    parameters:
      - $ref: "#/components/parameters/RetreatId"
      - $ref: "#/components/parameters/AssignmentId"
    delete:
      tags: [Assignments]
      summary: Remove assignment
      operationId: deleteAssignment
      responses:
        "204":
          description: Removed
        "404":
          $ref: "#/components/responses/NotFound"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    FirebaseBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Firebase Auth ID token
  parameters:
    RetreatId:
      name: retreatId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    TaskId:
      name: taskId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    AssignmentId:
      name: assignmentId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    SealedConfirmationToken:
      name: sealedConfirmationToken
      in: path
      required: true
      description: |
        URL-safe signed token (JWT or HMAC envelope) binding assignment id + volunteer id + expiry.
        Implemented in deployment; MUST NOT be derivable without server secret key.
      schema:
        type: string
        minLength: 32
...
  responses:
    Unauthorized:
      description: Missing or invalid Firebase token
    NotFound:
      description: Resource not found
  schemas:
...
    Volunteer:
      type: object
      required: [id, displayName, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        displayName:
          type: string
        email:
          type: string
          format: email
          nullable: true
        phone:
          type: string
          nullable: true
        otherDuties:
          type: string
          nullable: true
        notifyEmail:
          type: boolean
          description: |
            Mirrors `jewelheart_volunteers.notify_email`: allow transactional messaging on email addresses (SendGrid/other).
            Omitted legacy responses should be treated by clients as `true`.
        notifySms:
          type: boolean
          description: |
            Mirrors `jewelheart_volunteers.notify_sms`: allow transactional SMS where phone populated (Twilio/other).
            Omitted legacy responses should be treated by clients as `false`.
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Assignment:
      type: object
      required: [id, taskId, volunteerId, createdAt]
      properties:
        id:
          type: string
          format: uuid
        taskId:
          type: string
          format: uuid
        volunteerId:
          type: string
          format: uuid
        volunteer:
          $ref: "#/components/schemas/Volunteer"
          description: Optional embed
        createdAt:
          type: string
          format: date-time
    AssignmentCreate:
      type: object
      required: [volunteerId]
      properties:
        volunteerId:
          type: string
          format: uuid
...
    AssignmentConfirmationRespondRequest:
      type: object
      required: [intent]
      properties:
        intent:
          type: string
          enum: [committed, withdrawn]
          description: >
            `committed` acknowledges coverage; persistence optional.
            `withdrawn` SHOULD delete assignment (implementation default) pending staffing policy tweaks.
    AssignmentConfirmationRespondResult:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
        message:
          type: string
        assignmentRemoved:
          type: boolean
          description: True when withdrawn path deleted the JewelHeart assignment row.
...
security:
  - FirebaseBearer: []